k8s/deployment-to-daemonset/requirement.txt
fire==0.3.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2